feat(dashboard): consolidate 11 tabs to 7 with sub-views and activity panel
4-tier restructuring of the dashboard tab architecture:

Tier 1 — Tasks (65 LOC, read-only) absorbed into Planning/Goals tab
Tier 2 — Execution merged into Overview via Triage/Dispatch sub-tabs,
         common MonitorTone helpers extracted to common/monitor.ts
Tier 3 — Council merged into Board via Posts/Debates/Voting sub-tabs
Tier 4 — Activity converted from tab to global slide-out panel

Tab count: 11 → 7 (Observe: Overview, Agents | Coordinate: Board,
Planning | Command: Command, Ops, TRPG)

Deleted: tasks.ts, council.ts, execution.ts (kept as sub-component)
Legacy hash routes (#tasks, #execution, #council, #activity) redirect
gracefully via LEGACY_TAB_ALIASES in router.ts.
