+++ assets/dashboard/index.html
@@ -5,8 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>MASC Dashboard</title>
   <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-  <script type="module" crossorigin src="/dashboard/assets/index-Yz_5AW--.js"></script>
-  <link rel="stylesheet" crossorigin href="/dashboard/assets/index-CS6g0JBQ.css">
+  <script type="module" crossorigin src="/dashboard/assets/index-BNjkyIcC.js"></script>
+  <link rel="stylesheet" crossorigin href="/dashboard/assets/index-Btn0_Ys1.css">
 </head>
 <body>
   <div id="app"></div>
