+++ dashboard/src/components/execution.ts
@@ -4,6 +4,7 @@ import { html } from 'htm/preact'
 import { Card } from './common/card'
 import { StatusBadge } from './common/status-badge'
 import { TimeAgo } from './common/time-ago'
+import { type MonitorTone, toEpoch, toneRank, normalizeKey, limitText, taskPriorityValue, taskPriorityLabel } from './common/monitor'
 import { buildAgentMotion, type AgentMotionSnapshot } from './common/agent-motion'
 import { openAgentDetail } from './agent-detail'
 import { agents, boardPosts, keepers, messages, tasks } from '../store'
@@ -12,8 +13,6 @@ import { journal } from '../sse'
 
 const QUIET_EXECUTION_MS = 10 * 60 * 1000
 const STALE_EXECUTION_MS = 20 * 60 * 1000
-
-type MonitorTone = 'ok' | 'warn' | 'bad'
 type DispatchState = 'dispatchable' | 'drift' | 'loaded' | 'quiet' | 'offline'
 
 interface TaskExecutionRow {
@@ -60,42 +59,8 @@ type InterventionItem =
       agentRow: DispatchRow
     }
 
-function normalizeKey(value: string | null | undefined): string {
-  return (value ?? '').trim().toLowerCase()
-}
-
-function toEpoch(value: string | number | null | undefined): number {
-  if (value == null) return 0
-  const parsed = typeof value === 'number' ? value : Date.parse(value)
-  return Number.isNaN(parsed) ? 0 : parsed
-}
-
-function trimText(value: string | null | undefined, max = 96): string | null {
-  const normalized = (value ?? '').replace(/\s+/g, ' ').trim()
-  if (!normalized) return null
-  return normalized.length > max ? `${normalized.slice(0, max - 3)}...` : normalized
-}
-
-function toneRank(tone: MonitorTone): number {
-  switch (tone) {
-    case 'bad': return 2
-    case 'warn': return 1
-    default: return 0
-  }
-}
-
-function taskPriorityValue(priority?: number | null): number {
-  if (typeof priority !== 'number' || Number.isNaN(priority)) return 3
-  return priority
-}
-
-function taskPriorityLabel(priority?: number | null): string {
-  const value = taskPriorityValue(priority)
-  if (value <= 1) return 'P1'
-  if (value === 2) return 'P2'
-  if (value >= 4) return 'P4+'
-  return 'P3'
-}
+// trimText is an alias for limitText from common/monitor
+const trimText = limitText
 
 function taskStatusLabel(status: Task['status']): string {
   switch (status) {
